feat(module:OverlayTrigger): common logic relocation
