+++ components/tooltip/Tooltip.razor.cs
@@ -3,8 +3,6 @@ using AntDesign.Internal;
 using OneOf;
 using System.Threading.Tasks;
 using System.Linq;
-using AntDesign.JsInterop;
-using System.Text.Json;
 using Microsoft.AspNetCore.Components.Web;
 
 namespace AntDesign
@@ -22,9 +20,6 @@ namespace AntDesign
 
         [Parameter]
         public double MouseLeaveDelay { get; set; } = 0.1;
-
-        [Inject]
-        private DomEventService DomEventService { get; set; }
 
         public Tooltip()
         {
@@ -60,56 +55,6 @@ namespace AntDesign
             }
 
             await base.Hide(force);
-        }
-        protected override Task OnAfterRenderAsync(bool firstRender)
-        {
-            if (firstRender && Unbound != null)
-            {
-                Ref = RefBack.Current;
-                DomEventService.AddEventListener(Ref, "click", OnUnboundClick, true);
-                DomEventService.AddEventListener(Ref, "mouseover", OnUnboundMouseEnter, true);
-                DomEventService.AddEventListener(Ref, "mouseout", OnUnboundMouseLeave, true);
-                DomEventService.AddEventListener(Ref, "focusin", OnUnboundFocusIn, true);
-                DomEventService.AddEventListener(Ref, "focusout", OnUnboundFocusOut, true);
-                DomEventService.AddEventListener(Ref, "contextmenu", OnContextMenu, true, true);
-
-            }
-            return base.OnAfterRenderAsync(firstRender);
-        }
-
-        private void OnUnboundMouseEnter(JsonElement jsonElement) => OnTriggerMouseEnter();
-        private void OnUnboundMouseLeave(JsonElement jsonElement) => OnTriggerMouseLeave();
-        private void OnUnboundFocusIn(JsonElement jsonElement) => OnTriggerFocusIn();
-        private void OnUnboundFocusOut(JsonElement jsonElement) => OnTriggerFocusOut();
-
-        protected override void Dispose(bool disposing)
-        {
-            if (Unbound != null)
-            {
-                DomEventService.RemoveEventListerner<JsonElement>(Ref, "click", OnUnboundClick);
-                DomEventService.RemoveEventListerner<JsonElement>(Ref, "mouseover", OnUnboundMouseEnter);
-                DomEventService.RemoveEventListerner<JsonElement>(Ref, "mouseout", OnUnboundMouseLeave);
-                DomEventService.RemoveEventListerner<JsonElement>(Ref, "focusin", OnUnboundFocusIn);
-                DomEventService.RemoveEventListerner<JsonElement>(Ref, "focusout", OnUnboundFocusOut);
-                DomEventService.RemoveEventListerner<JsonElement>(Ref, "contextmenu", OnContextMenu);
-            }
-            base.Dispose(disposing);
-        }
-
-        private async void OnUnboundClick(JsonElement jsonElement)
-        {
-            var eventArgs = JsonSerializer.Deserialize<MouseEventArgs>(jsonElement.ToString(),
-                new JsonSerializerOptions { PropertyNameCaseInsensitive = true });
-
-            await base.OnClickDiv(eventArgs);
-        }
-
-        private async void OnContextMenu(JsonElement jsonElement)
-        {
-            var eventArgs = JsonSerializer.Deserialize<MouseEventArgs>(jsonElement.ToString(),
-                new JsonSerializerOptions { PropertyNameCaseInsensitive = true });
-
-            await base.OnTriggerContextmenu(eventArgs);
         }
     }
 }
